+++ client/src/pages/admin/Users/index.jsx
@@ -9,7 +9,7 @@ import ModalBS from '../../../components/Modal';
 import FormBS from '../../../components/Form';
 import * as Yup from "yup";
 import styles from "../PanelStyling/style.module.css";
-import { sweetAlert } from '../../../utils/sweetAlert2';
+import sweetAlert from '../../../utils/sweetAlert2';
 import Loader from '../../../components/Loader';
 import Input from '../../../components/InputFields';
 import useSocket from '../../../hooks/useSocket';
@@ -68,7 +68,7 @@ function Users()
 
     // Delete
     const drop = useCallback(async (_id) => {
-        sweetAlert("Are you sure?", "This action will permanently delete the record.", "confirm", "Yes, delete it!", null, async () => {
+        sweetAlert.confirm({ fn:async () => {
             try 
             {
                 await api.delete(`/user/${_id}`);
@@ -77,7 +77,7 @@ function Users()
             {
                 return error;
             } 
-        })
+        } });
     },[]);
 
     // Columns
